requirements.txt
requests
pyTelegramBotAPI
Pillow